improve error handling on start and wait of the container
If the container platform cannot provision the instance, the start()
will fail, which is a retriable error for the user.

At the end of the loop, if no success has happened, we need to throw an
error, so the DO triggers again the constructor if we haven't been
successful initialising the `container` property.